pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "corbett_cli"
authors = [
    {name = "Tom Corbett", email = "tom.corbettco@gmail.com"},
]
description = "Utilities for querying Google Sheets from Snowflake"
# readme = "README.rst"
requires-python = ">=3.9"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "requests==2.28.2",
    "click==8.1.3",
    "snowflake-connector-python==3.0",
    "python-dotenv==1.0.0",
    "pydantic==1.10.7",
    "pynamodb==5.4.1",
    "argon2-cffi==21.2.0"
]
dynamic = ["version"]

[project.optional-dependencies]
pdf = ["ReportLab>=1.2", "RXP"]
rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
corbett = "corbett.cli:cli"

[tool.setuptools.packages.find]
where = [
    "src"
]